hotel-scraper-clean/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
lxml==4.9.3
firecrawl-py==0.0.16
python-multipart==0.0.6
